feat(api): support selectors in window function `order_by` and `group_by`
